+++ src/main/kotlin/com/autonomousapps/AbstractExtension.kt
@@ -6,7 +6,9 @@ import org.gradle.api.file.RegularFile
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Provider
+import org.gradle.api.tasks.TaskProvider
 
+@Suppress("MemberVisibilityCanBePrivate")
 abstract class AbstractExtension(private val objects: ObjectFactory) {
 
   private val adviceOutput = objects.fileProperty()
@@ -27,5 +29,18 @@ abstract class AbstractExtension(private val objects: ObjectFactory) {
    */
   fun adviceOutput(): RegularFileProperty {
     return adviceOutput
+  }
+
+  internal var postProcessingTask: TaskProvider<out AbstractPostProcessingTask>? = null
+
+  /**
+   * Register your custom task that post-processes the [com.autonomousapps.advice.ComprehensiveAdvice]
+   * produced by this project.
+   */
+  fun registerPostProcessingTask(task: TaskProvider<out AbstractPostProcessingTask>) {
+    postProcessingTask = task
+    postProcessingTask.configure {
+      input.set(adviceOutput())
+    }
   }
 }